backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# LangChain for document loading and RAG
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# Document parsing dependencies (used by LangChain loaders)
pypdf>=3.17.0
python-pptx>=0.6.23
python-docx>=1.1.0
unstructured>=0.11.0

# LLM and embeddings
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
